Update 'abaco create' and 'abaco deploy' with additional options
The deployment workflow now supports '-D' which skips cacheing the actorID in a
local .ACTOR_ID file. It also detects and either errors or deals with a return
value of 'null', which usually indicates that the auth session has expired.

The deploy command also has the option to skip push to the container registry,
which speeds up redeploys where only environment variables or secrets are
being changed. This is invoked using '-K'.

The deploy command also supports a '-v' mode to report the underlying abaco
and docker commands that are being invoked by the workflow

The delay for finalizing the container registry push has been decreased to 2 sec