Support added for synchronous messages
The 'abaco run' command can now run an execution synchronously. This is
invoked using the '-x' option. At present, the only supported message
type and response is plaintext/json returning the execution log
from the execution. This will be extended in the future to support
bianry messages and streaming responses.