Fixed issue with needing to pass 'stateless' in an update operation
Abaco does not allow an actor to be updated to be stateless if it was
registered as stateful. But, the update operation currently requires
that the correct and current value for 'stateless' to be passed even
if no change is requested. We work around this by fetching the
actor's stateless key and simply sending that. In the future we may
permit toggling from stateless to stateful, but that remains to be
determined based on use cases and the Abaco road map."