Experimental support for automatic access token refresh
If the environment variable ABACO_ENABLE_AUTO_REFRESH == 1, all Abaco
commands attempt to refresh the session access token, preventing
Abaco management actions from failing due to accidentally expired
Oauth credentials. This function relies on the auth-tokens-refresh
command that ships with the Tapis CLI.

Alternative installations (or cases where the CLI is simply not
in the user PATH) are supported via specification of the TAPIS_CLI_PATH
environment variable. If the variable is set, but does not point to a
known directory, or is not set at all, the system resolves the
CLI path via dirname(which auth-tokens-refresh)."