Initial support for git integration w 'abaco init'
Projects created via 'abaco init' now initialize the new project as a git repo. This
is a forerunner to support automatically creating and connecting a git remote. Also
supported now is a base path (-B) for the new project and a text description (-d).
Finally, project templates do not need to be language-typed only so the help test
for the -l option was changed to reflect this fact.