Abaco deploy  honors all config vars and abaco init populates reactor.rc more fully.
Before this update, 'abaco deploy' was not able to honor REACTOR_USE_UID and REACTOR_PRIVILEGED
because it was checking for identity to the integer 1 rather than a string. The 'abaco init'
command did not populate the 'reactor.rc' file with entries for REACTOR_USE_UID, REACTOR_PRIVILEGED,
or REACTOR_STATELESS, making their existence unknown without reading source code. It also made no
attempt to bootstrap or allow user override of the container registry username / organization.