Added support for tagging with latest git commit hash
Bringing in a pattern observed in wild actor git repositories, if the
variable 'DOCKER_USE_COMMIT_HASH' is set to 1 in 'reactor.rc', and if
the 'git' binary is in '$PATH' and the current directory is an active
git repository with a least one commit, the short version of the
last commit is used to tag the built container 'image:<commit>'. If
the commit hash cannot be resolved, the :tag is set to ':latest'.

For users adopting a 'deploy-on-commit' workflow, this setting makes it
very easy to align generated image artifacts with specific source
code and configuration changes.