Added support for actor "token" property
The 'abaco init', 'abaco create', 'abaco update', and 'abaco deploy' commands now support the
'token' property for Actors, which governs whether an actor's executions are populated with
Tapis access tokens. The registration-time and project default behaviors can be tweaked via
combinations of the 'REACTOR_TOKENS' config variable and the '-A' command line flag.

This resolves issue #13